requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
flake8==7.0.0
inflection==0.5.1
install==1.3.5
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
model-bakery==1.18.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1